requirements.txt
dbt-core
dbt-duckdb
dlt[duckdb]
openai
psutil
pandas
numpy
flake8
